Polish router module structure
