+++ backend/app/routers/__init__.py
@@ -1,0 +1,9 @@
+"""Collection of FastAPI routers."""
+
+from . import users, payments, stats
+
+__all__ = [
+    "users",
+    "payments",
+    "stats",
+]
