+++ backend/tests/test_monthly_drinks.py
@@ -53,9 +53,10 @@ def test_monthly_drinks():
     ]
     db.add_all(events)
     db.commit()
+    user_id = user.id
     db.close()
 
-    resp = client.get(f"/users/{user.id}/monthly_drinks")
+    resp = client.get(f"/users/{user_id}/monthly_drinks")
     assert resp.status_code == 200
     data = resp.json()
     months = {row["month"]: row["count"] for row in data}
